refactor: merge messenger-types into messenger-internal
affects: @userlike/messenger-internal, @userlike/messenger
