+++ packages/messenger-internal/src/v1/util.ts
@@ -1,0 +1,4 @@
+import { State } from "./types";
+
+export const getUnreadMessageCount = (state: State): number =>
+  state.conversations.reduce((acc, c) => acc + c.unread.length, 0);
