+++ packages/messenger-internal/tsconfig.json
@@ -1,9 +1,4 @@
 {
-  "references": [
-    {
-      "path": "../messenger-types"
-    }
-  ],
   "extends": "../../config/tsconfig.es.json",
   "include": ["src/**/*.ts"],
   "compilerOptions": {
