--- packages/messenger-types/CHANGELOG.md
@@ -1,157 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [0.1.0-alpha.4](https://github.com/userlike/messenger/compare/v0.1.0-alpha.3...v0.1.0-alpha.4) (2020-08-25)
-
-**Note:** Version bump only for package @userlike/messenger-types
-
-
-
-
-
-# [0.1.0-alpha.3](https://github.com/userlike/messenger/compare/v0.1.0-alpha.2...v0.1.0-alpha.3) (2020-08-25)
-
-
-### Features
-
-* add state$ and getState ([881d8f1](https://github.com/userlike/messenger/commit/881d8f14c945035ec82dafeeb77f37207f5b4f4c))
-
-
-
-
-
-# [0.1.0-alpha.0](https://github.com/userlike/messenger/compare/v0.0.1-alpha.15...v0.1.0-alpha.0) (2020-08-06)
-
-
-### Features
-
-* add setContactInfo ([641c6cf](https://github.com/userlike/messenger/commit/641c6cfdfce1c6981a4699abe9844825791bb81c))
-
-
-
-
-
-## [0.0.1-alpha.16](https://github.com/userlike/messenger/compare/v0.0.1-alpha.15...v0.0.1-alpha.16) (2020-08-06)
-
-**Note:** Version bump only for package @userlike/messenger-types
-
-
-
-
-
-## [0.0.1-alpha.13](https://github.com/userlike/messenger/compare/v0.0.1-alpha.12...v0.0.1-alpha.13) (2020-07-20)
-
-**Note:** Version bump only for package @userlike/messenger-types
-
-
-
-
-
-## [0.0.1-alpha.12](https://github.com/userlike/messenger/compare/v0.0.1-alpha.11...v0.0.1-alpha.12) (2020-06-30)
-
-**Note:** Version bump only for package @userlike/messenger-types
-
-
-
-
-
-## [0.0.1-alpha.11](https://github.com/userlike/messenger/compare/v0.0.1-alpha.10...v0.0.1-alpha.11) (2020-06-29)
-
-**Note:** Version bump only for package @userlike/messenger-types
-
-
-
-
-
-## [0.0.1-alpha.10](https://github.com/userlike/messenger/compare/v0.0.1-alpha.9...v0.0.1-alpha.10) (2020-06-29)
-
-**Note:** Version bump only for package @userlike/messenger-types
-
-
-
-
-
-## [0.0.1-alpha.9](https://github.com/userlike/messenger/compare/v0.0.1-alpha.8...v0.0.1-alpha.9) (2020-06-29)
-
-**Note:** Version bump only for package @userlike/messenger-types
-
-
-
-
-
-## [0.0.1-alpha.8](https://github.com/userlike/messenger/compare/v0.0.1-alpha.7...v0.0.1-alpha.8) (2020-06-29)
-
-
-### Features
-
-* use a modular script loader, relying on widget key as a unique identifier ([0386673](https://github.com/userlike/messenger/commit/038667388ae491e984d72d89e09ae4d0cd4b2fc9))
-
-
-
-
-
-## [0.0.1-alpha.7](https://github.com/userlike/messenger/compare/v0.0.1-alpha.6...v0.0.1-alpha.7) (2020-06-23)
-
-
-### Bug Fixes
-
-* downgrade es target version, rm tsbuildinfo from dist ([c8eae24](https://github.com/userlike/messenger/commit/c8eae242cbef0e1790f50e32fcc6a8ac7a72205b))
-
-
-
-
-
-## [0.0.1-alpha.6](https://github.com/userlike/messenger/compare/v0.0.1-alpha.5...v0.0.1-alpha.6) (2020-06-23)
-
-
-### Features
-
-* add cjs output ([9e66508](https://github.com/userlike/messenger/commit/9e66508c759f77b9cfe9beca3aa42725ace09dff))
-
-
-
-
-
-## [0.0.1-alpha.5](https://github.com/userlike/messenger/compare/v0.0.1-alpha.4...v0.0.1-alpha.5) (2020-06-22)
-
-**Note:** Version bump only for package @userlike/messenger-types
-
-
-
-
-
-## [0.0.1-alpha.4](https://github.com/userlike/messenger/compare/v0.0.1-alpha.3...v0.0.1-alpha.4) (2020-06-22)
-
-**Note:** Version bump only for package @userlike/messenger-types
-
-
-
-
-
-## [0.0.1-alpha.3](https://github.com/userlike/messenger/compare/v0.0.1-alpha.2...v0.0.1-alpha.3) (2020-06-22)
-
-
-### Features
-
-* add baseUrl option ([c0d1cee](https://github.com/userlike/messenger/commit/c0d1ceed1501883455feca90f9ee8ec02ea72dbc))
-
-
-
-
-
-## [0.0.1-alpha.1](https://github.com/userlike/messenger/compare/v0.0.1-alpha.0...v0.0.1-alpha.1) (2020-06-21)
-
-**Note:** Version bump only for package @userlike/messenger-types
-
-
-
-
-
-## 0.0.1-alpha.0 (2020-06-21)
-
-
-### Features
-
-* first commit ([0af94b2](https://github.com/userlike/messenger/commit/0af94b21ff176a3069c516301559cc2d94d89261))
