--- packages/messenger-types/README.md
@@ -1,1 +1,0 @@
-See the [homepage]().
