--- packages/messenger-types/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@userlike/messenger-types",
-  "version": "0.1.0-alpha.4",
-  "main": "dist/cjs/index.js",
-  "module": "dist/es/index.js",
-  "types": "dist/es/index.d.ts",
-  "sideEffects": false,
-  "homepage": "https://github.com/userlike/messenger",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/userlike/messenger.git",
-    "directory": "packages/messenger-types"
-  },
-  "license": "MIT",
-  "files": [
-    "dist/es",
-    "dist/cjs"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "build:cjs": "tsc -b tsconfig.cjs.json",
-    "clean": "rm -rf dist",
-    "lint": "eslint src --ext .ts",
-    "test": "jest"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
