--- packages/messenger-types/src/index.ts
@@ -1,7 +1,0 @@
-import * as V1 from "./versions/v1";
-
-export * from "./ActionResult";
-export * from "./types";
-export * from "./versioning";
-
-export { V1 };
