--- packages/messenger-types/tsconfig.cjs.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../config/tsconfig.cjs.json",
-  "include": ["src/**/*.ts"],
-  "compilerOptions": {
-    "outDir": "dist/cjs",
-    "rootDir": "src"
-  }
-}
