--- packages/messenger-types/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../config/tsconfig.es.json",
-  "include": ["src/**/*.ts"],
-  "compilerOptions": {
-    "outDir": "dist/es",
-    "rootDir": "src"
-  }
-}
