+++ packages/messenger/tsconfig.cjs.json
@@ -2,9 +2,6 @@
   "references": [
     {
       "path": "../messenger-internal"
-    },
-    {
-      "path": "../messenger-types"
     }
   ],
   "extends": "../../config/tsconfig.cjs.json",
